serviços e controllers de Consultas

+++ WorkspaceReservas/Data/Dto/ConsultaDTO.cs
@@ -2,8 +2,8 @@
 {
     public class ConsultaDTO
     {
-        public int Id { get; set; }
-        public int MedicoIdFk { get; set; }
+        public long Id { get; set; }
+        public long MedicoIdFk { get; set; }
         public required string NomePaciente { get; set; }
         public DateTime DataHoraInicio { get; set; }
         public DateTime DataHoraFim { get; set; }
